refactor: SearchBarのロジックをuseSearchBarカスタムフックに分離
